backend/requirements.txt
pandas
yfinance
sqlalchemy
psycopg2-binary
python-dotenv
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
httpx